# apps/backend/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: DropDate API
  version: "0.1.0"
  description: |-
    Проста API, щоб перевірити статус сервера та дізнатись дату наступного релізу фільму/серіалу.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      description: Повертає простий статус, що сервер живий.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /next-release:
    get:
      summary: Наступна дата релізу
      description: >
        Повертає інформацію про найближчий реліз (серіали або фільми) з TMDB.
        Результати кешуються приблизно на 30 хвилин, тому повторні запити швидші.
      parameters:
        - in: query
          name: title
          schema:
            type: string
          required: true
          description: Назва фільму або серіалу (нечутлива до регістру). Обовʼязкова навіть якщо передаєте `tmdbId`.
        - in: query
          name: tmdbId
          schema:
            type: integer
          required: false
          description: Опціонально — конкретний TMDB ID для точного пошуку.
        - in: query
          name: mediaType
          schema:
            type: string
            enum: [movie, tv]
          required: false
          description: Тип тайтлу в TMDB (movie або tv).
      responses:
        '200':
          description: Вдало знайшли реліз
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseInfo'
        '400':
          description: Не передали title
        '404':
          description: Не знайшли реліз
  /suggest:
    get:
      summary: Підказки для пошуку
      description: Повертає короткий список варіантів із TMDB для автодоповнення.
      parameters:
        - in: query
          name: query
          schema:
            type: string
          required: true
          description: Рядок пошуку (мінімум 2 символи).
        - in: query
          name: limit
          schema:
            type: integer
            default: 5
          required: false
          description: Скільки результатів повернути (1-10).
      responses:
        '200':
          description: Короткі підказки з TMDB
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suggestion'
        '400':
          description: Некоректний рядок пошуку
  /trending:
    get:
      summary: Популярне зараз
      description: Повертає популярні фільми і серіали з TMDB.
      parameters:
        - in: query
          name: window
          schema:
            type: string
            enum: [day, week]
            default: day
          required: false
          description: Вікно трендів (day або week).
        - in: query
          name: limit
          schema:
            type: integer
            default: 18
          required: false
          description: Скільки результатів повернути для кожного типу.
      responses:
        '200':
          description: Списки популярних фільмів і серіалів
          content:
            application/json:
              schema:
                type: object
                properties:
                  movies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suggestion'
                  series:
                    type: array
                    items:
                      $ref: '#/components/schemas/Suggestion'
        '500':
          description: Помилка TMDB
  /bulk-next-release:
    post:
      summary: Пакетне оновлення релізів
      description: >
        Повертає масив результатів для списку тайтлів або TMDB ID.
        Кожен елемент відповідає одному вхідному `clientId` і містить `info` або `error`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [clientId]
                    properties:
                      clientId:
                        type: string
                        example: "saved:the-boys"
                      title:
                        type: string
                        example: "The Boys"
                      tmdbId:
                        type: integer
                        example: 76479
                      mediaType:
                        type: string
                        enum: [movie, tv]
              required: [items]
      responses:
        '200':
          description: Пакетна відповідь
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        clientId:
                          type: string
                        info:
                          $ref: '#/components/schemas/ReleaseInfo'
                        error:
                          type: string
        '400':
          description: Некоректний payload
components:
  schemas:
    ReleaseInfo:
      type: object
      properties:
        title:
          type: string
          example: "Dune: Part Two"
        type:
          type: string
          enum: [movie, series]
        nextRelease:
          type: string
          format: date-time
          example: 2024-03-01T00:00:00Z
        source:
          type: string
          example: tmdb-mock
        posterUrl:
          type: string
          format: uri
          example: "https://image.tmdb.org/t/p/w342/abc123.jpg"
        backdropUrl:
          type: string
          format: uri
          example: "https://image.tmdb.org/t/p/w780/def456.jpg"
        status:
          type: string
          enum: [upcoming, released, ended]
          example: upcoming
    Suggestion:
      type: object
      properties:
        id:
          type: integer
          example: 31917
        title:
          type: string
          example: "Dune: Part Two"
        mediaType:
          type: string
          enum: [movie, tv]
        year:
          type: string
          example: "2024"
        posterUrl:
          type: string
          format: uri
          example: "https://image.tmdb.org/t/p/w342/abc123.jpg"